--- dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/imports/TrackerImportScheduleEventJob.java
@@ -1,100 +1,0 @@
-/*
- * Copyright (c) 2004-2025, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.webapi.controller.tracker.imports;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.hisp.dhis.common.UID;
-import org.hisp.dhis.event.EventStatus;
-import org.hisp.dhis.scheduling.Job;
-import org.hisp.dhis.scheduling.JobConfiguration;
-import org.hisp.dhis.scheduling.JobProgress;
-import org.hisp.dhis.scheduling.JobType;
-import org.hisp.dhis.tracker.imports.TrackerEventScheduleParams;
-import org.hisp.dhis.tracker.imports.TrackerImportParams;
-import org.hisp.dhis.tracker.imports.TrackerImportService;
-import org.hisp.dhis.tracker.imports.TrackerImportStrategy;
-import org.hisp.dhis.tracker.imports.domain.Event;
-import org.hisp.dhis.tracker.imports.domain.MetadataIdentifier;
-import org.hisp.dhis.tracker.imports.domain.TrackerObjects;
-import org.hisp.dhis.user.CurrentUserUtil;
-import org.hisp.dhis.util.DateUtils;
-import org.springframework.stereotype.Component;
-
-/**
- * @author Zubair Asghar
- */
-@Component
-@RequiredArgsConstructor
-public class TrackerImportScheduleEventJob implements Job {
-  private final TrackerImportService trackerImportService;
-
-  @Override
-  public JobType getJobType() {
-    return JobType.TRACKER_IMPORT_EVENT_SCHEDULE_JOB;
-  }
-
-  @Override
-  public void execute(JobConfiguration config, JobProgress progress) {
-    TrackerEventScheduleParams scheduleParams =
-        (TrackerEventScheduleParams) config.getJobParameters();
-
-    TrackerObjects trackerObjects =
-        TrackerObjects.builder()
-            .events(
-                List.of(
-                    createEvent(
-                        scheduleParams.getEnrollment(),
-                        scheduleParams.getOrgUnit(),
-                        scheduleParams.getProgramStageUid(),
-                        scheduleParams.getAttributeOptionCombo(),
-                        scheduleParams.getScheduledAt())))
-            .build();
-
-    TrackerImportParams importParams =
-        TrackerImportParams.builder()
-            .skipRuleEngine(true)
-            .importStrategy(TrackerImportStrategy.CREATE)
-            .build();
-    trackerImportService.importTracker(importParams, trackerObjects);
-  }
-
-  private Event createEvent(
-      String enrollment, String orgUnit, String programStageUid, String aoc, String scheduledAt) {
-    return Event.builder()
-        .event(UID.generate())
-        .enrollment(UID.of(enrollment))
-        .orgUnit(MetadataIdentifier.ofUid(orgUnit))
-        .programStage(MetadataIdentifier.ofUid(programStageUid))
-        .attributeOptionCombo(MetadataIdentifier.ofUid(aoc))
-        .storedBy(CurrentUserUtil.getCurrentUsername())
-        .scheduledAt(DateUtils.instantFromDateAsString(scheduledAt))
-        .status(EventStatus.SCHEDULE)
-        .build();
-  }
-}
